feat: introduce Services and run_bot entry point
Extracts handler wiring from main.rs into lib::run_bot, generic over
Transport + LoginCredentials so integration tests can call the same
entry point with a fake transport. Services struct bundles
test-overridable deps (Clock, LlmClient, AviationClient, data_dir).
main.rs shrinks from 563 to 90 lines. Helper fns moved into library
modules (twitch_setup.rs, llm::build_llm_client, config::load_configuration).
