+++ src/llm/mod.rs
@@ -1,9 +1,12 @@
 pub mod ollama;
 pub mod openai;
 
+use std::sync::Arc;
+
 use async_trait::async_trait;
 use eyre::Result;
 use serde::{Deserialize, Serialize};
+use tracing::{debug, error, info};
 
 /// A message in a chat completion conversation.
 #[derive(Debug, Clone)]
@@ -86,4 +89,49 @@ pub trait LlmClient: Send + Sync {
         &self,
         request: ToolChatCompletionRequest,
     ) -> Result<ToolChatCompletionResponse>;
+}
+
+/// Build an [`LlmClient`] from optional AI config.
+///
+/// Returns `Ok(None)` when no AI config is provided or AI is disabled.
+/// Returns `Ok(Some(client))` when the client is successfully built.
+/// Returns `Err` only when the configuration is invalid (not when the backend is unreachable).
+pub fn build_llm_client(ai_config: Option<&crate::config::AiConfig>) -> Result<Option<Arc<dyn LlmClient>>> {
+    use secrecy::ExposeSecret as _;
+    use crate::config::AiBackend;
+
+    let Some(ai_cfg) = ai_config else {
+        debug!("AI not configured, AI command disabled");
+        return Ok(None);
+    };
+
+    let result = match ai_cfg.backend {
+        AiBackend::OpenAi => {
+            let api_key = ai_cfg
+                .api_key
+                .as_ref()
+                .expect("validated: openai backend has api_key");
+            openai::OpenAiClient::new(
+                api_key.expose_secret(),
+                &ai_cfg.model,
+                ai_cfg.base_url.as_deref(),
+            )
+            .map(|c| Arc::new(c) as Arc<dyn LlmClient>)
+        }
+        AiBackend::Ollama => {
+            ollama::OllamaClient::new(&ai_cfg.model, ai_cfg.base_url.as_deref())
+                .map(|c| Arc::new(c) as Arc<dyn LlmClient>)
+        }
+    };
+
+    match result {
+        Ok(client) => {
+            info!(backend = ?ai_cfg.backend, model = %ai_cfg.model, "LLM client initialized");
+            Ok(Some(client))
+        }
+        Err(e) => {
+            error!(error = ?e, "Failed to initialize LLM client");
+            Ok(None)
+        }
+    }
 }
